feat: implement soft delete functionality for Course entity
- Updated delete method to mark the Course as deleted by setting the deleted flag to true instead of performing a physical delete.

+++ src/main/java/apps/sarafrika/elimika/course/service/CourseService.java
@@ -2,7 +2,6 @@ package apps.sarafrika.elimika.course.service;
 
 import apps.sarafrika.elimika.course.dto.request.CreateCourseRequestDTO;
 import apps.sarafrika.elimika.course.dto.response.CourseResponseDTO;
-import apps.sarafrika.elimika.course.model.Course;
 import apps.sarafrika.elimika.shared.dto.ResponseDTO;
 import apps.sarafrika.elimika.shared.dto.ResponsePageableDTO;
 import org.springframework.data.domain.Pageable;
@@ -16,5 +15,5 @@ public interface CourseService {
 
     CourseResponseDTO update(Course course);
 
-    void delete(int id);
+    void delete(Long id);
 }
